move attach button to bottom left
