+++ quadratic-client/src/app/ui/components/AIContext.tsx
@@ -5,14 +5,13 @@ import {
 } from '@/app/atoms/aiAnalystAtom';
 import { events } from '@/app/events/events';
 import { sheets } from '@/app/grid/controller/Sheets';
-import { uploadFile } from '@/app/helpers/files';
+
 import { getTableNameFromPos } from '@/app/quadratic-core/quadratic_core';
 import type { CodeCell } from '@/app/shared/types/codeCell';
 import { defaultAIAnalystContext } from '@/app/ui/menus/AIAnalyst/const/defaultAIAnalystContext';
-import { AttachFileIcon, CloseIcon } from '@/shared/components/Icons';
+import { CloseIcon } from '@/shared/components/Icons';
 import { Button } from '@/shared/shadcn/ui/button';
 import { HoverCard, HoverCardContent, HoverCardTrigger } from '@/shared/shadcn/ui/hover-card';
-import { TooltipPopover } from '@/shared/shadcn/ui/tooltip';
 import { cn } from '@/shared/shadcn/utils';
 import {
   getDataBase64String,
@@ -20,7 +19,7 @@ import {
   isSupportedImageMimeType,
 } from 'quadratic-shared/ai/helpers/files.helper';
 import type { Context, FileContent, UserMessagePrompt } from 'quadratic-shared/typesAndSchemasAI';
-import { memo, useCallback, useEffect, useMemo, useState } from 'react';
+import { memo, useCallback, useEffect, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import { AIAnalystSelectContextMenu } from '../menus/AIAnalyst/AIAnalystSelectContextMenu';
 
@@ -30,25 +29,12 @@ type AIContextProps = {
   setContext?: React.Dispatch<React.SetStateAction<Context>>;
   files: FileContent[];
   setFiles: React.Dispatch<React.SetStateAction<FileContent[]>>;
-  handleFiles: (files: FileList | File[]) => void;
-  fileTypes: string[];
   editing: boolean;
   disabled: boolean;
   textareaRef: React.RefObject<HTMLTextAreaElement | null>;
 };
 export const AIContext = memo(
-  ({
-    initialContext,
-    context,
-    setContext,
-    files,
-    setFiles,
-    handleFiles,
-    fileTypes,
-    editing,
-    disabled,
-    textareaRef,
-  }: AIContextProps) => {
+  ({ initialContext, context, setContext, files, setFiles, editing, disabled, textareaRef }: AIContextProps) => {
     const loading = useRecoilValue(aiAnalystLoadingAtom);
     const messages = useRecoilValue(aiAnalystCurrentChatMessagesAtom);
     const messagesCount = useRecoilValue(aiAnalystCurrentChatMessagesCountAtom);
@@ -126,8 +112,6 @@ export const AIContext = memo(
             onClose={handleOnCloseSelectContextMenu}
           />
         )}
-
-        <AttachFileButton disabled={disabled} handleFiles={handleFiles} fileTypes={fileTypes} />
 
         {files.map((file, index) => (
           <FileContextPill
@@ -220,34 +204,6 @@ const FileContextPill = memo(({ disabled, file, onClick }: FileContextPillProps)
         <span className="block border-t border-border px-1 py-0.5 text-xs">{file.fileName}</span>
       </HoverCardContent>
     </HoverCard>
-  );
-});
-
-type AttachFileButtonProps = {
-  disabled: boolean;
-  handleFiles: (files: FileList | File[]) => void;
-  fileTypes: string[];
-};
-const AttachFileButton = memo(({ disabled, handleFiles, fileTypes }: AttachFileButtonProps) => {
-  const handleUploadFiles = useCallback(async () => {
-    const files = await uploadFile(fileTypes);
-    handleFiles(files);
-  }, [handleFiles, fileTypes]);
-
-  const label = useMemo(() => (fileTypes.includes('.pdf') ? 'Attach image or PDF' : 'Attach image'), [fileTypes]);
-
-  return (
-    <TooltipPopover label={label}>
-      <Button
-        size="icon-sm"
-        className="h-5 w-5 shadow-none"
-        variant="outline"
-        onClick={handleUploadFiles}
-        disabled={disabled}
-      >
-        <AttachFileIcon className="!h-5 !w-5 !text-sm" />
-      </Button>
-    </TooltipPopover>
   );
 });
